refactor: restructure endpoints
