--- api/app/api/api_v1/endpoints/reddit/__init__.py
@@ -1,3 +1,0 @@
-from . import info, label, sync
-
-__all__ = ["info", "label", "sync"]
