+++ api/app/api/api_v1/routes.py
@@ -1,11 +1,10 @@
 from fastapi import APIRouter
 
-from .endpoints import posts, predict, reddit
+from .endpoints import info, posts, predict, sync
 
 api_router = APIRouter()
 
 api_router.include_router(posts.router, prefix="/posts", tags=["posts"])
 api_router.include_router(predict.router, prefix="/predict", tags=["predict"])
-api_router.include_router(reddit.label.router, prefix="/reddit/label", tags=["reddit"])
-api_router.include_router(reddit.info.router, prefix="/reddit/info", tags=["reddit"])
-api_router.include_router(reddit.sync.router, prefix="/reddit/sync", tags=["reddit"])
+api_router.include_router(info.router, prefix="/info", tags=["info"])
+api_router.include_router(sync.router, prefix="/sync", tags=["sync"])
